Made the function of deleting and restoring the task

+++ server/database.py
@@ -1,9 +1,7 @@
 import motor.motor_asyncio
-import json
-from bson import json_util
+from fastapi import HTTPException
 
 from model import ToDo
-
 
 client = motor.motor_asyncio.AsyncIOMotorClient()
 
@@ -17,14 +15,18 @@ async def fetch_all_todos():
     return todos
 
 
+async def fetch_one_todo(title):
+    document = await main_list.find_one({"title": title})
+    return document
+
+
 async def create_todo(todo):
     todo_title = todo.dict()["title"]
     check_db = await main_list.find_one({"title": todo_title})
     if check_db:
-        return None
-    else:
-        result = await main_list.insert_one(todo.dict())
-        return result
+        raise HTTPException(409, "A task with this name already exists")
+    result = await main_list.insert_one(todo.dict())
+    return result
 
 
 async def update_todo(title, desc):
@@ -33,25 +35,19 @@ async def update_todo(title, desc):
     return document
 
 
-async def fetch_one_todo(title):
-    document = await main_list.find_one({"title": title})
-    return document
-
-
 async def remove_todo(title):
     document = await main_list.find_one({"title": title})
-    # json_data = json.loads(json.dumps(document, default=json_util.default))
-    # print(json_data, type(json_data))
-    # del json_data["_id"]
-    # del json_data["time"]
-    # await main_list.delete_one({"title": title})
-    # return dict(json_data)
-    await main_list.delete_one({"title": title})
-    return document
+    if document:
+        del document["time"]
+        result = ToDo(**document)
+        await main_list.delete_one({"title": title})
+        return result
+
 
 async def move_to_trash_bin(todo):
-    document = await trash_bin.insert_one(todo)
-    return document
+    document = await trash_bin.insert_one(todo.dict())
+    if document:
+        return todo
 
 
 async def fetch_trash_bin():
@@ -63,3 +59,9 @@ async def delete_all_todos(collection):
     await collection.delete_many({})
     return True
 
+
+async def permanently_remove_todo(title):
+    document = await trash_bin.find_one({"title": title})
+    if document:
+        await trash_bin.delete_one({"title": title})
+        return True
